+++ wsgid/commands/__init__.py
@@ -1,5 +1,6 @@
 
 
-import init
-import config
-import restart
+from init import *
+from config import *
+from restart import *
+from stop import *